pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = "A boilerplate template to kickstart new Python projects."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.1.25",
    "colorlog>=6.9.0",
    "gitpython>=3.1.46",
    "jira>=3.10.5",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "pygithub>=2.8.1",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "tomli-w>=1.2.0",
    "typer>=0.21.1",
]

[dependency-groups]
dev = [
    "absolufy-imports>=0.3.1",
    "bandit>=1.8.3",
    "codespell>=2.4.1",
    "mypy>=1.16.0",
    "pre-commit>=4.2.0",
    "pydantic[mypy]>=2.11.5",
    "pytest>=8.3.5",
    "pytest-cases>=3.8.6",
    "ruff>=0.11.12",
]
